Move distraction scripts to new namespace
